Few updates, simplified skill system, added some comments
