--- src/cardboardbox/garbagebin/D20RollerDefunct.java
@@ -1,21 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package cardboardbox.garbagebin;
-
-import cardboardbox.rollers.Dice;
-
-
-
-/**
- *
- * @author egand
- */
-public class D20RollerDefunct{
-//    public int roll(int die, int mod, int attr){
-//        Dice diceRoller = new Dice();
-//        return ((diceRoller.roll(die)) + mod + attr);
-//    }
-}
