--- src/cardboardbox/garbagebin/HcHitCalc.java
@@ -1,41 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package cardboardbox.garbagebin;
-
-import cardboardbox.CardboardBox;
-import java.util.ArrayList;
-import java.time.*;
-import java.util.Date;
-
-/**
- *
- * @author egand
- */
-public class HcHitCalc extends CardboardBox {
-
-    public ArrayList hits(ArrayList rolls, int attr) {
-//        ArrayList res = new ArrayList();
-//        ArrayList suc = new ArrayList();
-//        int fails = 0, neut = 0, hit = 0;
-//        suc.add(0, fails);suc.add(1, neut);suc.add(2, hit);
-//        res.add(0, rolls);
-//        int bp = 5 - attr;
-//        for (int i = 0; i < rolls.size(); i++) {
-//            int x = (int) rolls.get(i);
-//            if (x == 1) {
-//                suc.set(0, (++fails));
-//            } else if (x > bp) {
-//                suc.set(2, (++hit));
-//            } else {
-//                suc.set(1, (++neut));
-//            }
-//        }
-//        res.add(1, suc);
-//
-        return null;
-    }
-
-}
