--- src/cardboardbox/garbagebin/HeroicRollerDefunct.java
@@ -1,24 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package cardboardbox.garbagebin;
-
-import cardboardbox.rollers.Dice;
-import java.util.ArrayList;
-
-/**
- *
- * @author egand
- */
-public class HeroicRollerDefunct {
-//    public ArrayList roll(int num){
-//        ArrayList res = new ArrayList();
-//        Dice diceRoll = new Dice();
-//        for(int i = 0;i<num;i++){
-//            res.add((diceRoll.roll(6)));
-//        }
-//        return res;
-//    }
-}
