--- src/manager/test2.xml
@@ -1,116 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<character>
-  <Biography>
-    <name>Alessandro Draconis</name>
-    <age>42</age>
-    <job>Emperor</job>
-    <race>Human (Dragonblooded)</race>
-    <height>76.0</height>
-    <weight>221.0</weight>
-    <morality>Authoritarian, Humanist</morality>
-  </Biography>
-  <CharacterState>
-    <totExp>10000</totExp>
-    <currentExp>0</currentExp>
-    <spentExp>10000</spentExp>
-    <level>50</level>
-    <threshold>0</threshold>
-    <charClass>Bravado</charClass>
-    <featList>
-      <feat>
-        <name>Example</name>
-        <description>This Is An Example</description>
-        <cost>100</cost>
-        <category>Examples</category>
-      </feat>
-    </featList>
-  </CharacterState>
-  <HealthState>
-    <health>180</health>
-    <will>120</will>
-    <status>alive</status>
-  </HealthState>
-  <Skills>
-    <Charisma>
-      <hustle>0</hustle>
-      <intimidate>0</intimidate>
-      <speech>0</speech>
-      <perform>0</perform>
-      <rank>0</rank>
-    </Charisma>
-    <Combat>
-      <magic>0</magic>
-      <rune>0</rune>
-      <defense>0</defense>
-      <placeholder>0</placeholder>
-      <rank>0</rank>
-    </Combat>
-    <Constitution>
-      <travel>0</travel>
-      <ride>0</ride>
-      <survival>0</survival>
-      <recover>0</recover>
-      <rank>0</rank>
-    </Constitution>
-    <Dexterity>
-      <acrobatics>0</acrobatics>
-      <sneak>0</sneak>
-      <sleight>0</sleight>
-      <ply>0</ply>
-      <rank>0</rank>
-    </Dexterity>
-    <Intelligence>
-      <investigate>0</investigate>
-      <gadgetry>0</gadgetry>
-      <knowledge>0</knowledge>
-      <reason>0</reason>
-      <rank>0</rank>
-    </Intelligence>
-    <Melee>
-      <assault>0</assault>
-      <rush>0</rush>
-      <shift>0</shift>
-      <bait>0</bait>
-      <rank>0</rank>
-    </Melee>
-    <Ranged>
-      <fire>0</fire>
-      <hamper>0</hamper>
-      <disengage>0</disengage>
-      <mark>0</mark>
-      <rank>0</rank>
-    </Ranged>
-    <Strength>
-      <strongarm>0</strongarm>
-      <athletics>0</athletics>
-      <breakSkill>0</breakSkill>
-      <haul>0</haul>
-      <rank>0</rank>
-    </Strength>
-    <Wisdom>
-      <survey>5</survey>
-      <insight>9</insight>
-      <gather>3</gather>
-      <triage>7</triage>
-      <rank>1</rank>
-    </Wisdom>
-  </Skills>
-  <Spellbook>
-    <Physical>
-      <spell>
-        <id>1</id>
-        <name>Example Spell</name>
-        <description>This is an Example</description>
-        <cost>1000</cost>
-      </spell>
-    </Physical>
-    <Magical>
-      <spell>
-        <id>0</id>
-        <name>Example Spell 2</name>
-        <description>This is a Magic Example Spell</description>
-        <cost>0</cost>
-      </spell>
-    </Magical>
-  </Spellbook>
-</character>
