--- src/ui/HeroicForm.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<!--
-To change this license header, choose License Headers in Project Properties.
-To change this template file, choose Tools | Templates
-and open the template in the editor.
--->
-<html>
-    <head>
-        <title>Heroic Character Sheet Test</title>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    </head>
-    <BODY>
-        This is the applet:<P>
-    <APPLET code="src/importer/HeroicImporter.class" width="300" height="50">
-        <PARAM name="Message" value="Howdy, there!">
-        This page will be very boring if your 
-        browser doesn't understand Java.
-    </APPLET>
-</BODY>
-</html>
